Fixed float types wrongly declared as NSUInteger in ETColumnFragment protocol
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@6415 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/ETFragment.h
@@ -75,19 +75,19 @@ treated under various circumstances:
 Warning: this API will probably evolve a bit. */
 @protocol ETColumnFragment
 /** Sets the column width. */
-- (void) setWidth: (NSUInteger)width;
+- (void) setWidth: (float)width;
 /** Returns the column width. */
-- (NSUInteger) width;
+- (float) width;
 /** Sets the minimum width allowed and resizes the column if the current width 
 is inferior. */
-- (void) setMinWidth: (NSUInteger)width;
+- (void) setMinWidth: (float)width;
 /** Returns the minimum width allowed. */
-- (NSUInteger) minWidth; 
+- (float) minWidth; 
 /** Sets the maximum allowed width and resizes the column if the current width 
 is superior. */
-- (void) setMaxWidth: (NSUInteger)width;
+- (void) setMaxWidth: (float)width;
 /** Returns the maximum width allowed. */
-- (NSUInteger) maxWidth;
+- (float) maxWidth;
 /** Sets the resizing behavior:
 
 <deflist>
